requirements_corrected.txt
asgiref==3.8.1
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==5.1.4
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pillow==11.0.0
python-decouple==3.8
requests==2.32.3
rfc3986==1.5.0
setuptools==75.1.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wheel==0.44.0
